Turret Inventory UI
Added Inventory UI to the alpha scene with no function currently.

Inventory framework is to appear when in "Build Mode".

Slots are to store references to each turret type.

Clicking on a slot as if it's a button selects it as the currently selected turret.

Hovering a turret spawns a tooltip sprite that showcases what that turret does.